only write new files if they don't exist yet

+++ gliderExport.R
@@ -35,9 +35,12 @@ for (i in deployedGliders$Name){
       select(any_of(needVars)) %>%
       mutate(m_present_time = as.numeric(m_present_time))
     
+    #write new file if not already present
+    if (!file.exists(paste0("/echos/segments/", i, "/", seg, ".csv"))){
     write.csv(segdf, 
                 file = paste0("/echos/segments/", i, "/", seg, ".csv"), 
                 row.names = FALSE, col.names = TRUE, quote = FALSE)
+    }
     
   }
 }
